refactor(desktop): group smaller handlers into useMiscActions (F5.10)
Stats/config/clear-caches modals, move-to-folder (single + bulk),
search-from-sender, open-in-Gmail, save message/.eml, label suggestions,
and saved-query CRUD move into useMiscActions. Verbatim; state stays in
App and arrives via deps.

App.tsx 1,652 -> 1,465; useMiscActions 247. unit 61 / e2e 39 green, tsc +
build clean.
