+++ desktop/frontend/src/App.tsx
@@ -14,10 +14,6 @@ import {
   type AnalyzerRule,
 } from "./api";
 import type { ComposeInit } from "./Compose";
-import {
-  emailAddr,
-  cleanSubject,
-} from "./format";
 import { replyInit, forwardInit } from "./compose";
 import ModalsPrimary from "./ModalsPrimary";
 import ModalsSecondary from "./ModalsSecondary";
@@ -35,6 +31,7 @@ import { runCommand } from "./commandRunner";
 import { useActionPlan } from "./useActionPlan";
 import { handleKeyDown } from "./keydownHandler";
 import { useAiActions } from "./useAiActions";
+import { useMiscActions } from "./useMiscActions";
 import type { KeydownCtx } from "./keydownCtx";
 import { useBootstrap } from "./useBootstrap";
 import { useZoom } from "./useZoom";
@@ -703,199 +700,15 @@ export default function App() {
   });
 
 
-  const openStats = useCallback(async () => {
-    setStatsOpen(true);
-    try {
-      setStats(await backend.UsageStats());
-    } catch (e) {
-      setError(String(e));
-    }
-  }, []);
-
-  const openConfig = useCallback(async () => {
-    setConfigOpen(true);
-    try {
-      setConfigInfo(await backend.ConfigInfo());
-    } catch (e) {
-      setError(String(e));
-    }
-  }, []);
-
-  const clearCaches = useCallback(async () => {
-    try {
-      await backend.ClearCaches();
-      showToast("Caches cleared");
-    } catch (e) {
-      setError(String(e));
-    }
-  }, [showToast]);
-
-  // doMove applies a label and archives the message (Gmail "move to folder").
-  const doMove = useCallback(
-    async (id: string, name: string) => {
-      const label = name.trim();
-      if (!label) return;
-      setMoveFor(null);
-      try {
-        await backend.MoveToLabel(id, label);
-        // Advance the cursor to the next row (like the TUI / archive / trash)
-        // instead of leaving the selection blank.
-        removeFromList(id);
-        showToast(`Moved to ${label}`);
-      } catch (e) {
-        setError(String(e));
-      }
-    },
-    [removeFromList, showToast],
-  );
-
-  // doBulkMove moves every selected message to a folder (apply label + archive),
-  // the bulk form of doMove — mirrors the TUI's bulk "move to folder".
-  const doBulkMove = useCallback(
-    async (name: string) => {
-      const label = name.trim();
-      if (!label) return;
-      const ids = [...selected];
-      setBulkMove(false);
-      if (ids.length === 0) return;
-      const idSet = new Set(ids);
-      const removed = messages
-        .map((m, i) => ({ m, i }))
-        .filter(({ m }) => idSet.has(m.id));
-      setBusy(true);
-      setBulkProgress(`Moving ${ids.length}…`);
-      setError("");
-      try {
-        await backend.BulkMoveToLabel(ids, label);
-        setMessages((prev) => prev.filter((m) => !idSet.has(m.id)));
-        advanceAfterBulk(idSet);
-        setSelected(new Set());
-        pushUndo(`move ${ids.length}`, async () => {
-          await backend.BulkUnarchive(ids);
-          removed.forEach(({ m, i }) => insertMessage(m, i));
-        });
-        showToast(`Moved ${ids.length} to ${label}`);
-      } catch (e) {
-        setError(String(e));
-      } finally {
-        setBulkProgress("");
-        setBusy(false);
-      }
-    },
-    [
-      selected,
-      messages,
-      advanceAfterBulk,
-      pushUndo,
-      insertMessage,
-      showToast,
-    ],
-  );
-
-  // quickSearch runs a Gmail search derived from the current message (sender,
-  // recipient, or subject), mirroring the TUI's F / T / S shortcuts.
-  const quickSearch = useCallback(
-    (kind: "from" | "to" | "subject", d: MessageDetail) => {
-      let q = "";
-      if (kind === "from") q = `from:${emailAddr(d.from) || d.from}`;
-      else if (kind === "to") q = `to:${emailAddr(d.to) || d.to}`;
-      else q = `subject:${JSON.stringify(cleanSubject(d.subject))}`;
-      setQuery(q);
-      void load(q);
-    },
-    [load],
-  );
-
-  const openInGmail = useCallback((id: string) => {
-    void backend.OpenGmailWeb(id).catch((e) => setError(String(e)));
-  }, []);
-
-
-  const saveMessage = useCallback(
-    (id: string) => {
-      void backend
-        .SaveMessage(id)
-        .then((path) => showToast(`Saved to ${path}`))
-        .catch((e) => setError(String(e)));
-    },
-    [showToast],
-  );
-
-  const saveRawMessage = useCallback(
-    (id: string) => {
-      void backend
-        .SaveRawMessage(id)
-        .then((path) => showToast(`Saved .eml to ${path}`))
-        .catch((e) => setError(String(e)));
-    },
-    [showToast],
-  );
-
-  const openSuggest = useCallback(async (id: string) => {
-    setSuggestFor(id);
-    setSuggestions([]);
-    setLoadingSuggest(true);
-    setError("");
-    try {
-      setSuggestions(await backend.SuggestLabels(id));
-    } catch (e) {
-      setError(String(e));
-    } finally {
-      setLoadingSuggest(false);
-    }
-  }, []);
-
-  const applySuggestion = useCallback(
-    (name: string) => {
-      if (!suggestFor) return;
-      void backend
-        .ApplyLabelByName(suggestFor, name)
-        .then(() => showToast(`Applied "${name}"`))
-        .catch((e) => setError(String(e)));
-      setSuggestions((prev) => prev.filter((s) => s !== name));
-    },
-    [suggestFor, showToast],
-  );
-
-
-  const openQueries = useCallback(async () => {
-    setQueriesOpen(true);
-    try {
-      setSavedQueries(await backend.ListSavedQueries());
-    } catch (e) {
-      setError(String(e));
-    }
-  }, []);
-
-  const runQuery = useCallback(
-    (q: SavedQuery) => {
-      setQueriesOpen(false);
-      void backend.RecordQueryUse(q.id).catch(() => undefined);
-      setQuery(q.query);
-      void load(q.query);
-    },
-    [load],
-  );
-
-  const deleteQuery = useCallback(async (id: number) => {
-    try {
-      await backend.DeleteSavedQuery(id);
-      setSavedQueries(await backend.ListSavedQueries());
-    } catch (e) {
-      setError(String(e));
-    }
-  }, []);
-
-  const doSaveQuery = useCallback(() => {
-    const name = saveQueryName.trim();
-    if (!name || !activeQuery) return;
-    void backend
-      .SaveQuery(name, activeQuery)
-      .then(() => showToast(`Saved query "${name}"`))
-      .catch((e) => setError(String(e)));
-    setSaveQueryOpen(false);
-    setSaveQueryName("");
-  }, [saveQueryName, activeQuery, showToast]);
+  const {
+    openStats, openConfig, clearCaches, doMove, doBulkMove, quickSearch, openInGmail, saveMessage, saveRawMessage, openSuggest, applySuggestion, openQueries, runQuery, deleteQuery, doSaveQuery,
+  } = useMiscActions({
+    messages, selected, activeQuery, suggestFor, saveQueryName, showToast,
+    setError, load, removeFromList, insertMessage, advanceAfterBulk, pushUndo,
+    setBulkMove, setBulkProgress, setBusy, setConfigInfo, setConfigOpen, setLoadingSuggest,
+    setMessages, setMoveFor, setQueriesOpen, setQuery, setSavedQueries, setSelected,
+    setStats, setStatsOpen, setSuggestFor, setSuggestions, setSaveQueryOpen, setSaveQueryName,
+  });
 
   const {
     runActionPlan, runDeterministicRules, applyCategory, dispatchPromptCategory,
